e2e: node: cpumanager: push utilities in util.go
Push more utilities from cpu manager tests to util.go.
This time there is not a clear (or big enough)
them for this set of shared code blocks.

Trivial code movement, no intended behavioral changes.
