+++ test/e2e_node/cpu_manager_test.go
@@ -26,8 +26,6 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
-	"k8s.io/kubelet/pkg/types"
 	kubeletconfig "k8s.io/kubernetes/pkg/kubelet/apis/config"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager"
 	admissionapi "k8s.io/pod-security-admission/api"
@@ -180,52 +178,6 @@ func makeCPUManagerInitContainersPod(podName string, ctnAttributes []ctnAttribut
 			},
 		},
 	}
-}
-
-func deletePodSyncByName(ctx context.Context, f *framework.Framework, podName string) {
-	gp := int64(0)
-	delOpts := metav1.DeleteOptions{
-		GracePeriodSeconds: &gp,
-	}
-	e2epod.NewPodClient(f).DeleteSync(ctx, podName, delOpts, f.Timeouts.PodDelete)
-}
-
-func deletePods(ctx context.Context, f *framework.Framework, podNames []string) {
-	for _, podName := range podNames {
-		deletePodSyncByName(ctx, f, podName)
-	}
-}
-
-func getLocalNodeCPUDetails(ctx context.Context, f *framework.Framework) (cpuCapVal int64, cpuAllocVal int64, cpuResVal int64) {
-	localNodeCap := getLocalNode(ctx, f).Status.Capacity
-	cpuCap := localNodeCap[v1.ResourceCPU]
-	localNodeAlloc := getLocalNode(ctx, f).Status.Allocatable
-	cpuAlloc := localNodeAlloc[v1.ResourceCPU]
-	cpuRes := cpuCap.DeepCopy()
-	cpuRes.Sub(cpuAlloc)
-
-	// RoundUp reserved CPUs to get only integer cores.
-	cpuRes.RoundUp(0)
-
-	return cpuCap.Value(), cpuCap.Value() - cpuRes.Value(), cpuRes.Value()
-}
-
-func waitForContainerRemoval(ctx context.Context, containerName, podName, podNS string) {
-	rs, _, err := getCRIClient()
-	framework.ExpectNoError(err)
-	gomega.Eventually(ctx, func(ctx context.Context) bool {
-		containers, err := rs.ListContainers(ctx, &runtimeapi.ContainerFilter{
-			LabelSelector: map[string]string{
-				types.KubernetesPodNameLabel:       podName,
-				types.KubernetesPodNamespaceLabel:  podNS,
-				types.KubernetesContainerNameLabel: containerName,
-			},
-		})
-		if err != nil {
-			return false
-		}
-		return len(containers) == 0
-	}, 2*time.Minute, 1*time.Second).Should(gomega.BeTrueBecause("Containers were expected to be removed"))
 }
 
 type cpuManagerKubeletArguments struct {
